Add CrudComponent with routing integration and styles
