+++ src/app/pages/pages.routes.ts
@@ -1,11 +1,11 @@
 import { Routes } from '@angular/router';
 import { Documentation } from './documentation/documentation';
-import { Crud } from './crud/crud';
+import { CrudComponent } from './crud/crud.component';
 import { Empty } from './empty/empty';
 
 export default [
     { path: 'documentation', component: Documentation },
-    { path: 'crud', component: Crud },
+    { path: 'crud', component: CrudComponent },
     { path: 'empty', component: Empty },
     { path: '**', redirectTo: '/notfound' }
 ] as Routes;
